MSI: make `exp' a short.

+++ msi.h
@@ -26,7 +26,7 @@
 #include <assert.h>
 #include <stdint.h>
 
-static inline uint32_t msi(int exp, uint64_t hash, uint32_t i) {
+static inline uint32_t msi(short exp, uint64_t hash, uint32_t i) {
   assert(-1 < exp);
   assert(33 > exp);
   uint32_t w = (uint32_t)((UINT64_C(1) << exp) - 1);
